requirements.txt
django
djangorestframework
djangorestframework-simplejwt
psycopg2
gunicorn
django-heroku